chore: update 'captp' package to use AVA for testing
