+++ packages/captp/package.json
@@ -25,7 +25,7 @@
   },
   "scripts": {
     "build": "exit 0",
-    "test": "tape -r esm 'test/**/*.js'",
+    "test": "ava",
     "lint-fix": "eslint --fix '**/*.js'",
     "lint-check": "eslint 'lib/*.js'",
     "lint-fix-jessie": "eslint -c '.eslintrc-jessie.js' --fix '**/*.js'",
@@ -33,9 +33,7 @@
   },
   "devDependencies": {
     "@agoric/install-ses": "^0.2.0",
-    "tap-spec": "^5.0.0",
-    "tape": "^4.11.0",
-    "tape-promise": "^4.0.0"
+    "ava": "^3.11.1"
   },
   "dependencies": {
     "@agoric/eventual-send": "^0.9.3",
@@ -49,5 +47,10 @@
   },
   "publishConfig": {
     "access": "public"
+  },
+  "ava": {
+    "files": ["test/**/test-*.js"],
+    "require": ["esm"],
+    "timeout": "2m"
   }
 }
